Write generation stepper logic
The code for the Cell class needed changes to reflect what's really
going on. Instead of previous and current states, what there are really
are the current state, and a temporary state. The current state is the
actual cell's alive/dead state. The temp state is where the new state as
a result of stepping a generation will reside, and relies on the current
states of the cells around it. After all cells receive their new temp
states, their current states are then updated.

I upped the demo rows and cols to 50, just for kicks (and it's boring to
watch life unfold on a small space). I also added some controls for
advancing a generation and resetting the thing.