Layout the controls
I contained the grid and the controls in a flex container so they are
centered horizontally. Also the controls is a flex container so the
buttons are on the left and the generation counter is on the right, with
space between them. The button group is also a flex container so the
buttons align nicely vertically with the generation counter.