Fix computation for cell row and column
I should have divided the index with the number of rows, not the number
of columns. The previous code worked fine if the grid is a square, but
behaved errantly when there are more rows than there are columns, and
refused to run at all when there are more columns (while bombarding the
console with out-of-bounds errors).

Naturally I didn't notice this mistake until I tried a non-square grid
in the styling branch.