Refactor and document GameOfLifeService
I reviewed the code fore the Game of Life service and added docs to the
functions, as well as refactor the code.

I made a getter for the generation counter.

Since the cells are exposed to the grid component anyway, I let it
toggle cell states directly, instead of having a function in the service
that does the same. I removed said function.

I added a check in the public functions to test if the grid has already
been initialized. The private functions don't have such checks because
the public functions that call them already do it.